requirements.txt
Flask==1.1.2
Flask-SocketIO==5.0.1
gunicorn==19.3.0
eventlet==0.30.1
PyJWT==1.7.1
mysql-connector-python==8.0.22

Authlib==0.15.3
requests==2.23.0